requirements.txt
langchain
langchain-core

# openai
langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface_hub


python-dotenv
scikit-learn


streamlit